Shift workflowlist 'component' into an actual page using layouts
